moved tests into triplet_test.py
